I introduced the idea of a game loop
It really is not much, it just post a message every frame saying
"render" and providing a timestamp for the current world. I have
never really written a game loop before, I think it is interesting
because it basically introduces the notion of time as an event. The
game loop is an event, the event is the passage of time. Kind of
interesting.